Extract shared imgproxy signer into packages/imgproxy
Both cove-image and cove-item carried an identical copy of the imgproxy
HMAC-SHA256 signing core. Pull it into a shared packages/imgproxy module
(the repo's first shared Go module) along with the canonical 5-variant
catalog, so the signing algorithm and variant sizes live in one place.

- New module packages/imgproxy with signer + variant catalog + tests
- Both services depend on it via require + a local replace directive
- Dockerfiles rebuilt to use the repo root as build context so the
  shared module is reachable; service Dockerfile selected via file:
- Repo-root .dockerignore keeps the widened context lean
- CI: cove-image/cove-item builds use context: . + file:; new
  packages/imgproxy test job; packages/imgproxy/** added to path filters
