+++ .github/workflows/ci-services.yml
@@ -8,12 +8,14 @@ on:
       - 'services/cove-api/**'
       - 'services/cove-image/**'
       - 'services/cove-item/**'
+      - 'packages/imgproxy/**'
       - '.github/workflows/ci-services.yml'
   pull_request:
     paths:
       - 'services/cove-api/**'
       - 'services/cove-image/**'
       - 'services/cove-item/**'
+      - 'packages/imgproxy/**'
       - '.github/workflows/ci-services.yml'
   # Allows manually triggering a build+push from any branch.
   # Useful for bootstrapping GAR before the integration branch merges to main,
@@ -129,11 +131,14 @@ jobs:
         if: github.event_name == 'push' && github.ref == 'refs/heads/main' || github.event_name == 'workflow_dispatch'
         run: gcloud auth configure-docker us-central1-docker.pkg.dev --quiet
 
+      # context is the repo root so the shared packages/imgproxy module is in
+      # the build context; the service Dockerfile is selected via file:.
       - name: Build and push image (main / manual)
         if: github.event_name == 'push' && github.ref == 'refs/heads/main' || github.event_name == 'workflow_dispatch'
         uses: docker/build-push-action@v6
         with:
-          context: services/cove-image
+          context: .
+          file: services/cove-image/Dockerfile
           push: true
           tags: us-central1-docker.pkg.dev/cove-6a685/services/cove-image:sha-${{ github.sha }}
           build-args: COMMIT_SHA=${{ github.sha }}
@@ -144,7 +149,8 @@ jobs:
         if: github.event_name == 'pull_request'
         uses: docker/build-push-action@v6
         with:
-          context: services/cove-image
+          context: .
+          file: services/cove-image/Dockerfile
           push: false
           build-args: COMMIT_SHA=${{ github.sha }}
 
@@ -189,11 +195,14 @@ jobs:
         if: github.event_name == 'push' && github.ref == 'refs/heads/main' || github.event_name == 'workflow_dispatch'
         run: gcloud auth configure-docker us-central1-docker.pkg.dev --quiet
 
+      # context is the repo root so the shared packages/imgproxy module is in
+      # the build context; the service Dockerfile is selected via file:.
       - name: Build and push image (main / manual)
         if: github.event_name == 'push' && github.ref == 'refs/heads/main' || github.event_name == 'workflow_dispatch'
         uses: docker/build-push-action@v6
         with:
-          context: services/cove-item
+          context: .
+          file: services/cove-item/Dockerfile
           push: true
           tags: us-central1-docker.pkg.dev/cove-6a685/services/cove-item:sha-${{ github.sha }}
           build-args: COMMIT_SHA=${{ github.sha }}
@@ -204,15 +213,43 @@ jobs:
         if: github.event_name == 'pull_request'
         uses: docker/build-push-action@v6
         with:
-          context: services/cove-item
+          context: .
+          file: services/cove-item/Dockerfile
           push: false
           build-args: COMMIT_SHA=${{ github.sha }}
+
+  # ── packages/imgproxy ─────────────────────────────────────────────────────
+  packages-imgproxy:
+    name: packages/imgproxy
+    runs-on: ubuntu-latest
+
+    permissions:
+      contents: read
+
+    steps:
+      - name: Checkout code
+        uses: actions/checkout@v6
+
+      - name: Set up Go
+        uses: actions/setup-go@v5
+        with:
+          go-version-file: packages/imgproxy/go.mod
+          # No external deps → no go.sum, so dependency caching has nothing to key on.
+          cache: false
+
+      - name: Run tests
+        working-directory: packages/imgproxy
+        run: go test ./...
+
+      - name: Run vet
+        working-directory: packages/imgproxy
+        run: go vet ./...
 
   # ── bump-overlay-tags ─────────────────────────────────────────────────────
   bump-overlay-tags:
     name: Open homelab PR — bump overlay tags
     runs-on: ubuntu-latest
-    needs: [cove-api, cove-image, cove-item]
+    needs: [cove-api, cove-image, cove-item, packages-imgproxy]
     # Only runs after a successful image push — same gate as the push steps above.
     if: (github.event_name == 'push' && github.ref == 'refs/heads/main') || github.event_name == 'workflow_dispatch'
