+++ packages/imgproxy/go.mod
@@ -1,0 +1,3 @@
+module github.com/danicajiao/cove/packages/imgproxy
+
+go 1.25.0
