+++ services/cove-image/Dockerfile
@@ -1,6 +1,10 @@
 # Stage 1 — Build
 # govips requires CGo (libvips C bindings), so we use a Debian-based Go image
 # and install the libvips development headers.
+#
+# The build context is the repo root (set in CI) so the shared packages/imgproxy
+# module — referenced via a replace directive in go.mod — is available to copy
+# in before the build runs.
 FROM golang:1.25-bookworm AS builder
 
 RUN apt-get update \
@@ -9,21 +13,27 @@ RUN apt-get update \
 
 WORKDIR /build
 
-# Download dependencies first so this layer is cached between builds
-# that only change source files.
-COPY go.mod go.sum ./
+# Shared module first. The replace directive in services/cove-image/go.mod
+# points at ../../packages/imgproxy, so it must exist on disk before `go build`
+# resolves it. The container layout mirrors the repo so the relative path lines up.
+COPY packages/imgproxy/ ./packages/imgproxy/
+
+# Service module files next, so the dependency-download layer caches
+# independently of service source changes.
+COPY services/cove-image/go.mod services/cove-image/go.sum ./services/cove-image/
+WORKDIR /build/services/cove-image
 RUN go mod download
 
-COPY . .
+# Service source.
+COPY services/cove-image/ ./
 
-# COMMIT_SHA is injected by CI via --build-arg so the /health endpoint
-# can report the exact Git commit that is running.
+# COMMIT_SHA is injected by CI via --build-arg so /health reports the running commit.
 ARG COMMIT_SHA=dev
 
 # CGO_ENABLED=1 is required — govips links against the libvips C library.
 RUN CGO_ENABLED=1 GOOS=linux go build \
     -ldflags "-X main.commitSHA=${COMMIT_SHA}" \
-    -o cove-image \
+    -o /build/cove-image \
     ./cmd/image/
 
 # Stage 2 — Run
